Linked tasks to phases
Tasks now belong to a phase. To configure tasks that belong to a workflow but are not considered to be a process step, specify the workflow but do not provide neither a next_id nor a rejected_task_id. Also make sure that no other tasks point to it since will cause the workflow to get stuck.

Signed-off-by: Justin Jeffress <justin.jeffress@booktrope.com>