ParseWrapper and ParseController
Created an api for interacting with Parse.com via the rails app.

ParseWrapper is the library that augments the parse-ruby-client gem by adding functionality specific to our parse.com instance and data models. All interactions with the parse api should go through the wrapper. If you're writing parse queries outside of the wrapper you're doing it wrong.

ParseController is the gateway to the ParseWrapper. We need to keep page loads zippy which is why parse api calls should never occur in the show action of a controller. They should be lazy loaded from the view via an ajax request to the ParseController. This allows the calls to the parse api to be concurrent which will help server performance and wont slow down page renders while waiting for a response. It also protects our parse api keys since they wont need to be declared in javascript.

Signed-off-by: Justin Jeffress <justin.jeffress@booktrope.com>