Fixed revenue split form validation
Changed the name of the validator method to revenueSplitValidation which is more explicit than percentageSum.

Updated jQuery.format call to jQuery.validator.format since the former is deprecated.

Changed the method for the form back to PATCH since we are using an embedded form to update the :team_memberships that belong to an already existing :project.

Task 176

https://booktrope.acunote.com/projects/47888/tasks/176
Signed-off-by: Justin Jeffress <justin.jeffress@booktrope.com>