1099 Form Task 174 & updates to coffee script
Added 1099 form support. Data doesn't get saved to the database. Still need to determine how we will save the data in the long run.

Updated the revenue split form's javascript to show a current tally of the percents so the user can easily tell if they are under the revenue threshold.

https://booktrope.acunote.com/projects/47888/tasks/174
Signed-off-by: Justin Jeffress <justin.jeffress@booktrope.com>