Task 207 - Upload Cover Concept
This is a pretty basic file upload.  I tested the approach using a validate_attachment
with the default constant for ContentType_Document and a .docx file, but it did not
pass validation properly.  On my personal machine, the documents show up as
"application/zip; charset=binary" which does not match the validation parameters. This
may be a problem for type validations.

https://booktrope.acunote.com/projects/47888/tasks/207