Task 187 - Approve Cover Art
This required adding code to support the rejected tasks as well as fixes in the seeds
file to also populate the reject task id.  It's also unique in that it uses an attribute
accessor as a placeholder for a field which is not on the model, but is used to determine
whether or not certain actions need to take place.  E.g. If they approve, update the date,
if they do not, update the notes field and reject the current task.

https://booktrope.acunote.com/projects/47888/tasks/187