Task 179 KDP Update
Added the update functionality.  The majority of the values are stored inside a JSON field
called update_data.  This should be easy to process by looking at the update_type to determine
which fields will be stored in update_data.  The form uses a make-shift wizard using a bunch of
divs and using show/hide.

https://booktrope.acunote.com/projects/47888/tasks/179