Task 202 Request Image
Added the Request Image form, recording the requested links to a JSON field in the cover_concepts table.  I
chose to take the existing dynamic table code I'd used before and just modify it.  I had attempted to make it
more flexible but it was taking too much time, so we'll want to go back and re-attempt it at a later date.

https://booktrope.acunote.com/projects/47888/tasks/202