Modified Create Project form, set project type, current_tasks, and populate some fields in project index.
The view for Project creation is now limited to authors or Booktrope staff.  If the current user is
Booktrope Staff, they are presented with a list of authors to select from.  If the current user is an
Author, their name is shown in bold below the label.

Upon creation, the team_memberships is created including the specified author and uses the suggested
percentage based on the require_roles value for the author for the selected project type.  Afterwards,
the current_tasks are populated with the initial values based each workflow for the selected project
type.

I've added a few field values to the Assets page, though they probably should get moved into partials.

I've also added some values from the project to the Project index so it's not just showing a static
value anymore.