Task 236 - Amazon S3 Setup & Paperclip configuration
Added support for Amazon S3 for storage and configured in Paperclip.  Almost all of the assets that are stored in S3
are stored with restricted access that requires an authorization token.  All links to downloadable assets use a
redirect method which issues a link that expires within ten seconds.

I've also added more partials for the assets page and cleaned a few other references to assets up.

https://booktrope.acunote.com/projects/47888/tasks/236