Task 235 - Lazy Load Parse Analytics
Created a new controller that's a singular resource for Project.  I liked this approach
because it allowed me to keep this logic out of the projects controller and provided
a clean url to work with.  This also opens up the possiblity of hitting a separate HTML
view at the same URL if we wanted to go for a full screen render.

https://booktrope.acunote.com/projects/47888/tasks/235