Task 326 - Add Image: needs to be .jpg or .zip (if they request more than one image we zip them and upload one zip file.
Added Javascript validation for the front-end and updated the file validation on the back-end.
In the Assets page, if the upload was an image then an image will appear beside the download link.  If the file is a zip file, then only the download link will show.
I've also added the list of image requests to the assets page with this change.

https://booktrope.acunote.com/projects/47888/tasks/326