Task 304 - Conversion: load avatar for Users in NewTrope
Created a new rake task called avatar_import:import which will pull from db/seed-data/wp-avatars.csv which should be committed before being run.

If the URL is not Gravatar (teamtrope.com) then it will be imported and the url stored.
If the URL is Gravatar, a lookup is done to determine if the user has a Gravatar (and not the default image).  If the url is valid, it is stored.
The profile helper for the avatar_url has been modified so that if a user does not have an avatar but an avatar_url is configured the avatar_url will be used.

https://booktrope.acunote.com/projects/47888/tasks/304