Task 434 - Consistently getting an application error when attempting to upload PDF and ebook files.
Implemented S3 direct uploads which uses a jQuery upload plugin that uploads directly to S3 using CORS.  After the upload has been completed, a URL and metadata for the file is passed to the back end.  The back-end processes the file by looking at it directly in S3 and moves if from an upload path into the intended path for the file type & project.

The first implemetation was done for Final Manuscript files.

https://booktrope.acunote.com/projects/47888/tasks/434