Task 513 - Fix Javascript bugs, ensure dynamic forms are working in Firefox
Used namespacing for some globals that needed to be set inside of a document ready event.
Rewrote the way the code checked for the root Teamtrope namespace object.
No longer double-initializing the acceptMember helper which seemed to cause issues.
Fixed a event handler's function declaration because FF didn't like it, even though Chrome was fine with it.

https://booktrope.acunote.com/projects/47888/tasks/513