Task 528 - Add new fields to New Project form
Added imprint, page_count, has_internal_illustrations, previously_published, and special_text_treatment fields to form.
Modified the intro paragraph with updated verbiage.
Fixed the notification for the new project creation so it gets sent.
Added new fields to the notification email and improved the genre so it lists all of the selected genres.
Fixed up formatting a bit to improve the display of radio buttons encapsulated in ul/li sets.

https://booktrope.acunote.com/projects/47888/tasks/528