Task 561 - Create an "author's books" grid that shows all books for a given author.
Modified the projects helper to support more methods for viewing the projects grid.
Added a author= search which uses the @nickname from OldTrope
Modified the user model to populate and update the nickname based on the user_nicename instead of the login from OldTrope.
Added an index to nickname on the users table to help speed lookups.

https://booktrope.acunote.com/projects/47888/tasks/561