Task 564 - Change Grid to include Book Cover Image thumbnail as 2nd column
Added a new cover_preview to the CoverTemplates model.  I chose to perform the thumbnail operation offline
since this is not a critical field.  Right now it's just a rake task but it could be modified as a resque
task in the future.

To enable, a job will need to be added to the Schedule in heroku so it runs teamtrope:update_cover_previews
on a regular basis.  (Every hour?)

https://booktrope.acunote.com/projects/47888/tasks/564