Task 502 - Add new Final Cover Approval step
Added a new approval step after the final covers is approved.
When a final cover is uploaded, an email is sent to Shari Ryan.
During the Approval step, the cover can be approved or denied.  In both cases notes can be stored.
If the cover is approved, it will move on to Design Complete which is what uploading the covers did previously.
The assets tab will indicate the current state, any notes (if they exist) and the last update of the covers.

https://booktrope.acunote.com/projects/47888/tasks/502