Task 540 - Allow Team to Update Cover at any time in the process until final covers are submitted.
Added the 'Cover Concept' task to the unlocked tasks associated with the 'Approve Cover Art' and
'Final Covers' tasks so that covers can be uploaded through the process.

What was not specified in the ticket is for the current task to skip back to the Approve Cover Art
step when a new concept is uploaded.  That will require a bit more work, so I'm going to request
feedback on the ticket.

https://booktrope.acunote.com/projects/47888/tasks/540