Task 617 - Remove "Marketing Release Date" form from the Project / Add Target Marketing Launch Date to Proofread Complete form
Added a migration which will remove the Marketing Release Date task and dependencies.
Remove all references to the Marketing Release Date.
Added jQueryUI support, but only included the Datepicker module.
Added a new radio option and a datepicker input for the Target Release Date on the Proofread Complete form.
Added a new FakeInput class that can be used for an input field that doesn't related to a given model.

https://booktrope.acunote.com/projects/47888/tasks/617