Task 617 - Add Target Marketing Launch Date to Proofread Complete form
Updated to note the date needs to be at least three weeks in the future.
Added a minDate parameter to the datepicker so the user doesn't have the option to use a previous date.
Added validation to check for manual user input to ensure the date is at least three weeks in the future.

https://booktrope.acunote.com/projects/47888/tasks/617