Task 664 - Add ability to set the genre
Added a new tab that provides the ability to select the genre for the book.  There are a few
little hacks in place because we're still keeping the framework to allow for multiple
selection through a one-to-many project -> book_genres relationship but want to only allow
for a single option for now.

https://booktrope.acunote.com/projects/47888/tasks/664