Fixed JSON Parser Error
There was a JSON parser error since we were still passing layout_approval_issue_list in the layout_attributes. If we submitted the form for the first time and there wan't a value in layout_approval_issue_list we would try to set it to "" and the JSON Parser complained.

Signed-off-by: Justin Jeffress <justin.jeffress@booktrope.com>