TTR-119 Hellosign callback map type to outsiders
If a hellosign document was created outside of teamtrope, we attempt to look up its type based on the subject, since hellosign doesn't pass the template that it's based on back to us. Unfortunately there is also no way to look it via the API.

In the future we will only rely on teamtrope for creating signing requests.

Signed-off-by: Justin Jeffress <justin.jeffress@booktrope.com>