Changed start_year to 2009
Rolledback to publish book after publishing a book and the date was still present.

I did notice that there is a publication_date column on the Project which appears to not be in use. We might want to consider removing it as part of this story or create a story to remove the column. Either way it's a new migration that drops the column.

Signed-off-by: Justin Jeffress <justin.jeffress@booktrope.com>