Bug fix
Added a missing referce to f which is the builder.

Without it the form was rendering the html without  project[layout_attributes] in the name which caused it to get ignored since it was not a layout attribute on the project.

It didn't throw an unpermitted parameters execption either since it was not part of the required parameters :project

Signed-off-by: Justin Jeffress <justin.jeffress@booktrope.com>