TTR-144 - Rewrote the float conversion to its own method.
I had never run into problems before with Nil class values which for some
reason support the .to_f method.  They not however support the .gsub()
method so an error would be thrown in the value for that column in the CSV
file did not exist.  I wrote a small helper method in the class to deal
with these sorts of conversions.