TTR-169 - Submitter does not receive notification for Print Corner submissions
Found a bug in the print_corners_request where the @current_user global was not
set so the send_submitter option could not send to the submitter.

I've added some logging so that if this condition exists again in the future
it will log an error so we should receive a notification from Logentries.

https://teamtrope.atlassian.net/browse/TTR-169