TTR-170 - Print Corners scaffold returns error page
I can't say for sure why, but one of the print corners records does
not return a user, and therefore the code would blow up when trying
to grab the nickname for a nil object when generating a profile link.

I can't find this record, but was able to code around the issue.

https://teamtrope.atlassian.net/browse/TTR-170