Retry failed jobs in the resque
This task is done but we will want to refactor it when we integrate the author agreemet, designer agreement, etc into teamtrope so we can:

a. re-run a subset of failed jobs
b. tailor the the validations to the specific failed job. (for example the project id is not required for an author agreement)

Signed-off-by: Justin Jeffress <justin.jeffress@booktrope.com>