Rake task report see description for details
This a report that lists projects that have 1 or more signing request in the failure queue. These projects are here because they need a legal_name which can be set using the edit screen for the project.

There are two report types, csv and urls. The csv report lists the project_ids and urls. The url report simply lists all the urls.

Signed-off-by: Justin Jeffress <justin.jeffress@booktrope.com>