Initial Commit of Report
Please do not approve this since it's not 100% complete.

There are two remaining tasks but I wanted to get some eyes on the code before I check it all in. The two remaining tasks are to compare the KDP Sales records with what we have in teamtrope which is just a database query and an update of the ChannelReportItem flag for kdp_select.

The second task is to update the nook retail page link for books that we find by looking them up using the search link (barnesandnoble.com/s/#{bnid}.

Signed-off-by: Justin Jeffress <justin.jeffress@booktrope.com>