Initial Commit of Report Email
This is not finalized but I wanted some eyes on it before hand.

Includes a few new instance methods on our models and some delegation of some functions from TeamMembership to User.

Project now has a function that returns the book_type in a human readable format, so I updated the ProjectMailer code to use that since the code was exactly the same.

User now has an instance method last_name_first that converts the display_name into last, first.

TeamMembership now delegates display_name and last_name_first to User.

Added a migration to rename the ParseBooks object_id column to parse_id since accessing the data via a query caused rails to throw a warning about against modifying object_id. Ours was overwriting theirs with ours from the db. It should be harmless but rather than stifle the warning message I renamed the column in the db.

Updated the parse_import rake task to use the new column name.

Signed-off-by: Justin Jeffress <justin.jeffress@booktrope.com>