TTR-237 - Potential Rights Back Form
Added a new form where an Imprint Manager or Booktrope staff can submit a request to return the book rights to the author.

What makes this unique is the requirement that the visibility of the form is dependent upon two factors:

1. A flag must first be set on the Project to enable the form
2. The user's login must be listed specifically in order to view the form

To configure the users, you must add an environment variable (or value in application.yml for your environment) with the name RIGHTS_BACK_REQUEST_ACCOUNTS, listing the login names in the value with only spaces between the values.

This adds a new check to the Task class which looks for a method named with the prefix being the task name (lowercase, underscores instead of spaces) with the suffix being _is_visible?.  (Example: Task is named "Rights Back Request", method should be named "rights_back_request_is_visibile?")  If the method exists it can perform necessary checks to determine if the Taba and related Task should be visible to the user.

https://teamtrope.atlassian.net/browse/TTR-237