Refactor Renamed proofed to proofread_final
This refactor touches a lot of files. I tried to be as thorough as possible I even updated the document import rake task (one line of code) and the test fixtures (which we don't use) to reflect the new name.

The only places where I left the word 'proofed' in tact were the rights back request, strings where it made sense, and proofed_word_count.

For testing we need to test submitting from the project and from the manuscript scaffolds.

Also ensure that the download links in the notification emails and in the project as well as the scaffold point to the correct file.

Signed-off-by: Justin Jeffress <justin.jeffress@booktrope.com>