Upload eBook from netgalley/ebook if none exists
I'm using the success callback to set which upload successfully updated via javascript and checking that in the create.js.erb for published_files.

The amazon partial now includes a new optional prefix local variable that will prefix the id of the uploader so it can be used multiple times on the same form.

In the future we may want to make the call to the partial also do the jquery bindings so we can move away from separate files.

Signed-off-by: Justin Jeffress <justin.jeffress@booktrope.com>