Added new Submit To Layout Form
This commit removes two steps from the workflow (proofread complete and choose style) and combines them into one.

The code has been left in place (partials and controller code) so we can roll back easily if necessary (without code changes/unmerging) we can remove the unecessary code if we deem it necessary to do so. :)

Signed-off-by: Justin Jeffress <justin.jeffress@booktrope.com>