Don't show published books in meta/scribd export
To mark a book so it wont appear in the meta spreadsheet and scribd, we need to first submit the rights back form which only certain booktrope staff members have access to. Once marked they will no longer appear in the above lists.

This change only blocks from the meta spreadsheet, and scribd exports. If we want to block them form the project we can add that to.

Signed-off-by: Justin Jeffress <justin.jeffress@booktrope.com>